Add hint_utils + fix format on execute_hint
