+++ src/hints/mod.rs
@@ -1,1 +1,2 @@
 pub mod execute_hint;
+pub mod hint_utils;
